crawl_vnexpress/requirements.txt
requests
tqdm
matplotlib
python-slugify
beautifulsoup4
lxml